+++ src/main/java/view/MessageResolver.java
@@ -1,6 +1,10 @@
 package view;
 
-import domain.*;
+import domain.board.ChessBoard;
+import domain.piece.Piece;
+import domain.board.File;
+import domain.position.Position;
+import domain.board.Rank;
 import view.mapper.PieceMapper;
 
 public class MessageResolver {
