+++ src/test/java/domain/TurnTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.board.ChessBoard;
 import fixture.PositionFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
