+++ src/test/java/domain/board/FileTest.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.board;
 
+import domain.board.File;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
