+++ src/test/java/domain/board/RankTest.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.board;
 
+import domain.board.Rank;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
